Remove UserStorage 'secret' logic
