+++ library/tiqr/Tiqr/UserStorage/Abstract.php
@@ -20,8 +20,6 @@
 use Psr\Log\LoggerInterface;
 
 require_once 'Tiqr/UserStorage/Interface.php';
-require_once 'Tiqr/UserStorage/Encryption.php';
-require_once 'Tiqr/UserSecretStorage.php';
 
 /**
  * Abstract base implementation for user storage. 
@@ -32,74 +30,11 @@ require_once 'Tiqr/UserSecretStorage.php';
  */
 abstract class Tiqr_UserStorage_Abstract implements Tiqr_UserStorage_Interface
 {
-    protected $_encryption;
-
-    protected $_userSecretStorage;
-
     protected $logger;
 
-    public function __construct($config, LoggerInterface $logger, $secretconfig = array())
+    public function __construct($config, LoggerInterface $logger)
     {
         $this->logger = $logger;
-        $type = isset($config['encryption']['type']) ? $config['encryption']['type'] : 'dummy';
-        $options = isset($config['encryption']) ? $config['encryption'] : array();
-        $this->_encryption = Tiqr_UserStorage_Encryption::getEncryption($logger, $type, $options);
-
-        if (count($secretconfig)) {
-            $this->_userSecretStorage = Tiqr_UserSecretStorage::getSecretStorage($secretconfig['type'], $logger, $secretconfig);
-        } else {
-            $this->_userSecretStorage = Tiqr_UserSecretStorage::getSecretStorage($config['type'], $logger, $config);
-        }
-    }
-
-    /**
-     * Returns the encryption instance.
-     */
-    protected function _getEncryption()
-    {
-        return $this->_encryption;
-    }
-
-    /**
-     * Get the user's secret
-     * @param String $userId
-     * @return String The user's secret
-     */
-    protected function _getEncryptedSecret($userId)
-    {
-        return $this->_userSecretStorage->getUserSecret($userId);
-    }
-
-    /**
-     * Store a secret for a user.
-     * @param String $userId
-     * @param String $secret
-     */
-    protected function _setEncryptedSecret($userId, $secret)
-    {
-        $this->_userSecretStorage->setUserSecret($userId, $secret);
-    }
-
-    /**
-     * Get the user's secret
-     * @param String $userId
-     * @return String The user's secret
-     */
-    public final function getSecret($userId)
-    {
-        $encryptedSecret = $this->_getEncryptedSecret($userId);
-        return $this->_getEncryption()->decrypt($encryptedSecret);
-    }
-
-    /**
-     * Store a secret for a user.
-     * @param String $userId
-     * @param String $secret
-     */
-    public final function setSecret($userId, $secret)
-    {
-        $encryptedSecret = $this->_getEncryption()->encrypt($secret);
-        $this->_setEncryptedSecret($userId, $encryptedSecret);
     }
 
     /**
