+++ library/tiqr/Tiqr/UserStorage/File.php
@@ -19,6 +19,7 @@
 
 use Psr\Log\LoggerInterface;
 
+require_once 'Tiqr/UserStorage/FileTrait.php';
 require_once 'Tiqr/UserStorage/GenericStore.php';
 
 /**
@@ -30,56 +31,18 @@ require_once 'Tiqr/UserStorage/GenericStore.php';
  */
 class Tiqr_UserStorage_File extends Tiqr_UserStorage_GenericStore
 {
-    protected $_path;
+    use FileTrait;
+
+    protected $path;
 
     /**
      * Create an instance
      * @param $config
      */
-    public function __construct($config, LoggerInterface $logger, $secretconfig = array())
+    public function __construct($config, LoggerInterface $logger)
     {
-        parent::__construct($config, $logger, $secretconfig);
-        $this->_path = $config["path"];
-    }
-
-    /**
-     * This function takes care of actually saving the user data to a JSON file.
-     * @param String $userId
-     * @param array $data
-     */
-    protected function _saveUser($userId, $data)
-    {
-        if (file_put_contents($this->getPath().$userId.".json", json_encode($data)) === false) {
-            $this->logger->error('Unable to save the user to user storage (file storage)');
-            return false;
-        }
-        return true;
-    }
-  
-    /**
-     * This function takes care of loading the user data from a JSON file.
-     * @param String $userId
-     * @return false if the data is not present, or an array containing the data.
-     */
-    protected function _loadUser($userId, $failIfNotFound = TRUE)
-    {
-        $fileName = $this->getPath().$userId.".json";
-
-        $data = NULL;
-        if (file_exists($fileName)) { 
-            $data = json_decode(file_get_contents($this->getPath().$userId.".json"), true);
-        }
-
-        if ($data === NULL) {
-            if ($failIfNotFound) {
-                throw new Exception('Error loading data for user: ' . var_export($userId, TRUE));
-            } else {
-                $this->logger->error('Error loading data for user from user storage (file storage)');
-                return false;
-            }
-        } else {
-            return $data;
-        }
+        parent::__construct($config, $logger);
+        $this->path = $config["path"];
     }
 
     /**
@@ -95,15 +58,4 @@ class Tiqr_UserStorage_File extends Tiqr_UserStorage_GenericStore
             $this->logger->error('Unable to remove the user from user storage (file storage)');
         }
     }
-
-    /**
-     * Retrieve the path where the json files are stored.
-     * @return String
-     */
-    public function getPath()
-    {
-         if (substr($this->_path, -1)!="/") return $this->_path."/";
-         return $this->_path;
-    }
-    
 }
